% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncran_close}
\alias{ncran_close}
\title{Close a connection.}
\usage{
ncran_close(ncid)
}
\arguments{
\item{ncid}{file connection provided by `ncran_open`}
}
\description{
Close a connection.
}
\examples{
f1 <- "S2008001.L3b_DAY_CHL.nc"
f_l3b <- system.file("extdata", "oceandata", f1, package = "ncran")
con <- ncran_open(f_l3b)
ncran_close(con)
}
