% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncran_inq_dims}
\alias{ncran_inq_dims}
\title{Dimension inquiry}
\usage{
ncran_inq_dims(grpid)
}
\arguments{
\item{grpid}{open connection}
}
\description{
Dimension inquiry
}
\examples{
f1 <- "S2008001.L3b_DAY_CHL.nc"
f_l3b <- system.file("extdata", "oceandata", f1, package = "ncran")
con <- ncran_open(f_l3b)
groupids <- ncran_inq_grps(con)
ncran_inq_dims(groupids[1])
ncran_close(con)
}
