% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncran_open}
\alias{ncran_open}
\title{Open a connection.}
\usage{
ncran_open(dsn)
}
\arguments{
\item{dsn}{data source name, i.e. file path or server path}
}
\description{
Pass  a character data source name to the internal nc_open library function.
}
\examples{
f1 <- "S2008001.L3b_DAY_CHL.nc"
f_l3b <- system.file("extdata", "oceandata", f1, package = "ncran")
con <-  ncran_open(f_l3b)
ncran_close(con)
f2 <- "S2008001.L3m_DAY_CHL_chlor_a_9km.nc"
f_l3m <- system.file("extdata", "oceandata", f2, package = "ncran")
con <-  ncran_open(f_l3m)
ncran_close(con)
#u_cst <- "http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdQSwind3da"
# Rnc_open(u_cst)
}
